+++ specula/processing_objects/dynamic_iir_filter.py
@@ -1,7 +1,6 @@
 from specula.processing_objects.iir_filter import IirFilter
 from specula.base_processing_obj import InputDesc
 from specula.data_objects.iir_filter_data import IirFilterData
-from specula.data_objects.simul_params import SimulParams
 from specula.connections import InputValue
 from specula.base_value import BaseValue
 from specula import cpuArray
@@ -15,8 +14,7 @@ class DynamicIirFilter(IirFilter):
     
     Parameters
     ----------
-    simul_params : SimulParams
-        Simulation parameters containing time step information
+
     iir_filter_data : IirFilterData
         Filter coefficients (numerator and denominator)
     delay : float [1], optional
@@ -38,7 +36,6 @@ class DynamicIirFilter(IirFilter):
     """
 
     def __init__(self,
-                 simul_params: SimulParams,
                  iir_filter_data: IirFilterData,
                  delay: float = 0,
                  integration: bool = True,
@@ -46,7 +43,6 @@ class DynamicIirFilter(IirFilter):
                  precision=None):
 
         super().__init__(
-            simul_params=simul_params,
             iir_filter_data=iir_filter_data,
             delay=delay,
             integration=integration,