+++ specula/processing_objects/integrator.py
@@ -3,12 +3,10 @@ from specula.processing_objects.iir_filter import IirFilter
 from specula.base_processing_obj import InputDesc, OutputDesc
 from specula.base_value import BaseValue
 from specula.data_objects.iir_filter_data import IirFilterData
-from specula.data_objects.simul_params import SimulParams
 
 
 class Integrator(IirFilter):
     def __init__(self,
-                 simul_params: SimulParams,
                  int_gain: list,
                  ff: list=None,
                  n_modes: list=None, # list[int]
@@ -52,7 +50,7 @@ class Integrator(IirFilter):
                                                target_device_idx=target_device_idx)
 
         # Initialize IirFilter object
-        super().__init__(simul_params, iir_filter_data, delay=delay, integration=integration,
+        super().__init__(iir_filter_data, delay=delay, integration=integration,
                          target_device_idx=target_device_idx, precision=precision)
 
     @classmethod